refactor(chatView): move subagent orchestration into SubagentController
The per-root tree refresh, history preview, follow-up and interrupt flows
owned six private fields and ~340 lines inside ChatViewProvider. Move them
whole into a vscode-free SubagentController (ProviderManagementDeps
precedent: runtime + currentRootSession() + onChange() injected), leaving
the provider with constructor subscriptions, five handleMessage cases,
postState/loadImage reads and dispose. The moved bodies are line-for-line
identical modulo mechanical renames; check and the 50-test gate pass.
